[REF] mail,im_livechat: move avatarUrl to channel
