+++ addons/im_livechat/static/src/embed/common/discuss_channel_model_patch.js
@@ -38,6 +38,12 @@ const discussChannelPatch = {
         }
         return super.displayName;
     },
+    get avatarUrl() {
+        if (this.channel_type === "livechat") {
+            return this.livechat_operator_id.avatarUrl;
+        }
+        return super.avatarUrl;
+    },
     get hasAttachmentPanel() {
         return this.channel_type !== "livechat" && super.hasAttachmentPanel;
     },
